ARM: dts: msm: Change SSC_CX and SSC_PX to proxy regulator type
Currently, the HLOS PIL driver votes for the SSC_CX and the SSC_PX
regulators while booting sensors and removes the votes only after
it shuts down.
Once the sensors subsystem is up and sends back a clock/power
ready interrupt, remove the HLOS vote as the assumption is that
the sensors subsystem would have voted for the regulators itself.

Change-Id: I00a41f77fcdcdc03379ce67986b2e428f67cf4dd
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>