msm: clock-thulium: Add the check_enable_bit flag to clock structures
Currently, during clk handoff, we determine if the clock is ON
in hardware by looking at the status bit. This might not always
work, now with the new votable smmu clock support. The votable
bits could be used to turn on multiple clocks and just looking
at the status bit for the individual clocks might be misleading.
Add the check_enable_flag to the impacted clock structures to
only check the enable bit for the HLOS vote.

Change-Id: I45dec5c964069200af84e2dd977f0055197994ba
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>