power: qpnp-fg: force BCL HPM during charge
The BCL device has 3 different power modes: Low Power Mode (LPM), Medium
Power Mode (MPM), and High Power Mode (HPM), in which it can switch
between. These different power modes corresponds to how frequent the BCL
does vbat and ibat conversions. The trigger to switch between these
modes is the battery current.

However, during charging or when an input is present, the battery
currrent can be very low. The BCL can go into MPM or LPM, which can
compromise the resolution of the BCL monitoring.

Force HPM when the device is charging in order to resolve this issue.

CRs-Fixed: 793296
Change-Id: I0b01f1749011bac3ba771a22efde9b6ba59e7ba9
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>