clk: msm: gdsc: Correct the check to see if GDSC is enabled
Currently, we only look at the status bit of the GDSC register to
determine if the GDSC is enabled or not. The regulator framework
uses this info to determine whether to call the regulator enable
callback. This works almost all the time but with the new votable
SMMU GDS support on Thulium, although the PWR_ON_MASK bit of the
GDS is set, HLOS might not have actually voted for it.
Check both the GDSC SW_COLLAPSE and PWR_ON_MASK to decide if the
GDSC is on/off.

Change-Id: I247f17dd469588dca0b98448ac6f19cec5b9a0c5
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>