msm: kgsl: Enable use of retention flops in the CP
Set a chicken bit to enable the use of retention flops in the CP
on the output of the RAM blocks. The read data on the read data
port is not being retained until the next read which seems to
result in the CP trying to access an unmapped address and causing
pagefaults. The chicken bit uses flops to capture and hold the
read data in case the RAMs cannot perform the function correctly.

Change-Id: I416cf99fab74605a94ab501c82a88c8b378aa16d
Signed-off-by: Shrenuj Bansal <shrenujb@codeaurora.org>