power: qpnp-fg: debounce soc empty in software
A sudden current spike may cause soc to be reported as 0 if the fuel
gauge does its vbat reading during the spike. Fix this by adding a
debounce time of 1.5 seconds, a full fuel gauge cycle, in order to make
sure a single spike cannot cause a failure.

Also, change some debug prints so that this condition is easier to debug.

CRs-Fixed: 799665
Change-Id: I02a15cdf02c5a331952c395fbf773cf2ae04ee7b
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>