clk: msm: gdsc: Add delay after enabling/disabling a GDSC
The GDS controller asserts/de-asserts the status bit soon after
it receives a power on/off request from a master. The controller
then takes around 8 xo cycles to start its internal state machine
and update the status bit. During this time, the status bit does
not reflect the true status of the core.
Add a delay of 1 us between writing to the SW_COLLAPSE bit and
polling the status bit.

Change-Id: I2662352a55ecae37975fcab15321ab7eefafe9c8
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>