msm: kgsl: Enable flush of shared cachelines every 10ms period
Enabling and disabling the UCHE timer based flush could cause the
UCHE to hang if it is already in the middle of a flush. OpenCL
needs the cache flushes for long running compute jobs so that
the CPU eventually gets visbility.

10ms has been determined as the appropriate interval for the timer
for performance and functionality. In addition, protect this
register so userspace cannot enable/disable the timer or
change the interval.

Change-Id: Ia3e005bea51475a6a0a6fb2f83db96fa852ae16c
Signed-off-by: Shrenuj Bansal <shrenujb@codeaurora.org>