ARM: dts: msm: Add bus master flag to camera dtsi CPP section
Some of the new platforms have CPP bus master. CPP driver
has to distinguish between the platforms which have CPP
bus master and those which don't have. A bus master flag
is added specially for this purpose in camera dtsi files.
If this flag is set to '1', bus master presents. If this
flag is set to '0' or it is missing then corresponding
platform doesn't have its own CPP bus master.

Change-Id: Ibd353cfedb2e4e518a2448dae094f061928ef144
Signed-off-by: Krishnankutty Kolathappilly <kkolatha@codeaurora.org>