clk: msm: clock-local2: Check the branch clk enable bit during handoff
Currently, during clk handoff, if a clock does not have the
check_enable_bit flag set, we determine if the clock is ON in
hardware by looking at the status bit. There might be cases
when the clock is enabled due to other HW reasons (debugger)
or due to incorrect bootchain configuration. Hence, instead
of determing that the clock is on right then, also check if
the enable bit is on and print a warning if it isnt.

Change-Id: I8c4ed90805826086a1f8b638b5c7af8a084fedd8
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>