ARM: dts: msm: Add clock properties to graphics GDSC node on MSM Thulium
The core and peripheral memories needed for graphics are not
enabled by default on Thulium. Add the bimc_gfx and gpu_gx_gfx3d
clock entries to the GDSC node so that the clock domain memories
can be turned on after enabling the GDSC and turned off before
disabling it.

Change-Id: I7ebf0e3732ba8cee2970cb2d0ab482b08bb4c50b
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>