power: smb135x-charger: fix USB type detection
Most of the platforms have D+/D- lines routed to USB PHY and the
charger. And almost all the PHYs pull down the D+/D- lines unless
it is set to active mode. This grounding causes issues with
source detection and results in detecting SDP even when it is not.

Fix this by setting ALLOW_DETECTION property to inform the PHY to
remove the pull downs. Once done, rerun APSD and inform of the correct
results via the TYPE and PRESENT property.
The reruns are necessary only if an SDP is detected.

To rerun APSD use this following procedure
1. Disable APSD
2. Set the allowable USB voltage to 9V.
3. Enable APSD
4. Set the allowable USB voltage to 5V-9V.
Step 4 causes APSD to run again and issue a src_detect interrupt.

Change-Id: I756bb34704be0fb5b85bcde74d22e8953f1c0ca2
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>