clk: msm: clock-thulium: Keep a few GCC clocks enabled always
The mmagic_video gds_hw_ctl clk_on signal is not connected to enable
the gcc_mmss_sys_noc_axi_clk as the hardware understanding is that
the clock is always kept running. There is also true in the case of
the aggre0_noc GDSC and the gcc_aggre0_noc_at_clk.
Add the always_on flag to these clock structures for this to happen.

Change-Id: I4d2fb93aa1221cae91dd5d810318783a364e2558
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>