msm: kgsl: Refactor the ioctl handler
The old ioctl code used absolute ioctl command values for matching
which meant that we were forced to maintain a match between
the kernel and userspace. In addition we blindly copied data to
and from the user using the size of the user specified ioctl
command meaning that 1) we had to be able to handle enormous
data sizes and 2) the kernel data size could never be greater
than the user data size (again - exact match needed). All this
meant that we had to maintain a lot of legacy ioctl codes and
add unused padding to new structs and generally trash.

Refactor the ioctl handler to remove absolute matches and to stop
trusting the user space so much. Use the kernel definition for
the ioctl command to define how much data space to allocate and
copy only that much into the struct (or the user size, whichever
is smaller). Move the tables and ioctl code into their own files,
add some helpers to copy the memory back and forth and untwist
the default and compat ioctl paths so they don't jump back and
forth between each other.

From now on we can now expand the size of a struct for an ioctl
on the kernel side without creating a new ioctl command code and
we can be way more tolerant of bogus incoming commands without having
a lot of extra check code laying about.

Change-Id: Ic0dedbadb89d50e9a7d5265c5fa9321fdee9a503
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>