soc: qcom: subsys-pil-tz: Conditionally disable regulators in proxy_unvote
The PIL driver votes for regulators on behalf of the subsystem
before loading it and bringing it out of reset. The subsystem,
on booting up, votes for the regulators from its end and sends
an interrupt back to PIL. At this point, PIL removes its voltage/
current vote and disables the regulators.
Add support to leave these regulators on while removing the
voltage/current votes when the qcom,allow-proxy-reg-enable flag
is set.

Change-Id: I31b9b84fb5c4999be8fa93c1cb69e90c9d13e4f3
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>