msm: kgsl: Emulate PWR performance counters on 5XX
The PWR performance counters do not physically exist on 5XX but
we need to provide performance counters that behave in a similar
fashion. Generic RBBM performance counters can be used to
count the events so swizzle the PWR perfcounters to use the RBBM
counters under the covers. This makes for more generic code and
and keeps the performance counter from automatically restoring the
contents of the counter after a power collapse.

Change-Id: Ic0dedbad09bfb8c32ebceaf182fcc3d636c0e81e
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>