msm: mdss: remove histogram collection timeout in software
Histogram collection will now be a blocking call, till either
of the following happen:
 1. Histogram data gets collected
 2. Suspend occurs causing a cancel on the blocking call
 3. No updates occur and notify timer says no update
 4. Some signal unblocks it

This is to be prevent too many timed out errors and unnecessary
mode switch from user mode to kernel mode on multiple histogram
read attempts.

The ECANCELED error being returned to userspace is expected to
be handled like a feedback to not call histogram till reenabled.
However a histogram stop call is still expected for disabling
histogram completely before the re-enablement.

Change-Id: I12fc0d56177f7ca64df77c10aa9a8340f054f889
Signed-off-by: Krishna Chaitanya Parimi <cparimi@codeaurora.org>