msm: kgsl: Clean up soft fault detect initialization
Most newer targets dont *NEED* soft fault detection but we keep it
around for fun. Move the initialization into its own function and
set a private bit if the initalization was successful.  This
consolidates bits and pieces from other functions and will make it
easier for us to featurize soft fault detection later.

Change-Id: Ic0dedbadd96ac571da57e8f6d388ecc1a4c0716f
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>