USB: hid: hid-core: Enable autosuspend feature for HID devices
Currently autosuspend feature for HID devices is disabled by default
in the kernel due to issues with suspend/resume issues with some of
the HID devices available in the market. On automotive platform atmel
HID device is connected on USB4 port and not entering to low power mode.

Enabling this autosuspend feature for HID devices as this is
the requirement to support auto suspend on ADP platform.

Change-Id: I18d3cb68468657d531b028da9bd8b0eb95e0802b
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>