msm: kgsl: Call the correct ioctl handler in kgsl_ioctl_helper()
Fix a very silly mistake in kgsl_ioctl_helper(). We were calling the
support function from the global kgsl_ioctl_funcs array instead of
from the passed in array.  This breaks in compat mode since program
flow was going directly to the 64 bit handlers which were completely
misinterpeting the structs.

Change-Id: Ic0dedbad08ad2ab66028c23aa772ac7d0cf5dfad
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>