msm: mhi: add MHI state machine
The Modem Host Interface (MHI) device driver supports
clients to send control and data packets such as
IP data packets, control messages and Diagnostic
data between the Host and the device. It follows the
MHI specification to transfer data.

MHI State Machine is a module in MHI device driver,
which is responsible for MHI state management and exposes
API to enable state transition of MHI client.

Change-Id: I46bcf1860ca6d6259a3a758424507bc535890cfd
Signed-off-by: Nadine Toledano <nadinet@codeaurora.org>