msm: mdss: set valid_roi as true when ctl paths are initialized
Mixer setup can be called from multiple contexts and not all may call
mixer set roi which sets valid_roi flag. However with partial update
there is a requirement that for a given ctl path, roi is not valid
because there is no update on that mixer. In such cases, valid_roi will
be false and we should ignore setting up the mixer. So to cover all
cases, initialize valid_roi to true when ctl path is setup first time.

Change-Id: I85ae4ce8711f0fcdf771d18b4896a8f61398ead5
Signed-off-by: Ujwal Patel <ujwalp@codeaurora.org>