diag: Forward commands to the diag clients registered on apps
Diag fails to  forward the commands registered on apps to
the user space clients since the proc id in the command
registration table is compared with the wrong proc id for
apps. Added a change to check the correct proc id for apps
while forwarding the command from kernel to userspace.

Change-Id: I4baac98fb752061b7621a6f725ce586bacd58a30
Signed-off-by: Sreelakshmi Gownipalli <sgownipa@codeaurora.org>