msm: vidc: Stop publishing frequency tables to devfreq
The frequency tables contains the min and max bandwidth values for each
bus. The governor might provide any value in the range of those two values.
However, if the device publishes a frequency table, devfreq expects the
governor to pick a enumerated in freq table, which doesn't really
suit some of the governors that we use.

The latter behaviour from devfreq is undesirable as it spams the log
"Couldn't update frequency transition information" each time we attempt
to scale buses.

As such, just stop publishing the table to devfreq and manually clamp
the governor's frequency vote in the .target() callback.

Change-Id: I167e1a7a43e643da501c452a31f320e2733cd94c
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>