msm: qpnp-power-on: change the usage of mutex while setting debounce delay
Currently, mutex lock of PON input is held while configuring the
debounce delay. However, when there are no input configurations
specified, debounce delay cannot be configured because of the
non-availability of input device and hence its mutex lock.

Fix it by using the mutex lock only when the input configuration
is available so that the debounce delay can be configured even
when there is no input configuration.

CRs-Fixed: 815913
Change-Id: Id4aaeceaae1a3c93dcdf962b7e23b187ebac966d
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>