ARM: dts: msm: Remove USB AB vote for MDM9640
AB vote is not required for bus clients such as USB.
This vote is used when client knows exactly what
is the expected bandwidth and bandwidth is guaranteed.
IB vote is sufficient to guarantee bus latencies for USB.
Any AB vote needs are met by IPA voting in high tput cases.

CRs-Fixed: 819015
Change-Id: Ie93a9150c257d3e1ef1786920b0381fcb78faaab
Signed-off-by: Devdutt Patnaik <dpatnaik@codeaurora.org>