msm: vidc : report PC_PREP failure as SYS_ERROR
When prepare for power collapse command fails, firmware is in
unusable state. This will make the driver unusable and all the
future sessions may fail till FW is unloaded and reloaded. This
commit reports this failure as SYS_ERROR resulting in unloading
the driver and firmware so next session can be opened successfully.

Change-Id: I849c2f8e98621bdff6ca4dddfd859a916e16dcb3
Signed-off-by: Ashray Kulkarni <ashrayk@codeaurora.org>