msm: camera: isp: Implementing synchoranization for dual vfe
In case of dual vfe, both vfe hardware is used for same camera
sensor. There are two tasklets serving irq's from each vfe. This
causes synchronization issues if either one of tasklet is delayed.
To fix this, implement synchoranization for dual vfe by using
right vfe to configure the image buffer for both vfe's and
Stats buffer to configure when we get the buffer for both vfe's.

Change-Id: If090bb7526c8d4f544f78e6a9378ebd382253940
Signed-off-by: Senthil Kumar Rajagopal <skrajago@codeaurora.org>
Signed-off-by: Harsh Shah <harshs@codeaurora.org>