msm: ipa: fix incorrect pipe number comparison
In ipa_tag_aggr_force_close, pipe number -1 as input indicates that
aggregation force close needs to be applied on all of the pipes.
Due to a comparison between an int and and unsigned int, the option
of -1 is not being exercised.
Similar instances which share this comparison were fixed as well.

Change-Id: Ib2a92ddd91dac74ea3ed44633191a73b6385e91d
Signed-off-by: Sivan Reinstein <sivanr@codeaurora.org>
Acked-by: David Arinzon <darinzon@qti.qualcomm.com>