msm: mdss: notify histogram is read ready using sysfs
Histogram calls for read need not be blocking in driver. A
sysfs entry can provide information on when the histogram
is ready to be read. The following are the changes related
to userspace protocol for getting histogram data:

1. Histogram read readiness will be notified with hist_event
   sysfs node to userspace for an enabled logical display,
   enabled through a MSMFB_START_HISTOGRAM ioctl
2. A read can be performed after the notification with the
   MSMFB_HISTOGRAM ioctl and it will be non blocking
3. The notification will happen when all histogram blocks of
   a particular logical display have completed collection
4. The change keeps histogram collection to still support
   a single client. Multiple clients cannot be present for
   reading of histogram data.

The change adds the display block field into the histogram
collection info struct. This field will be populated during
call to enable histogram and will help in identifying the
histogram blocks that are attached to that particular logical
display.

The change also removes the cancel histogram code during
suspend or no update scenario as the wait is now on sysfs.

Change-Id: I69d73e31018dc255fa51573146cddaec5d7af1a2
Signed-off-by: Krishna Chaitanya Parimi <cparimi@codeaurora.org>