msm: mdss: Use col_state rather than col_en during canceling hist
When histogram is canceled upon suspend or no_update event, the
histogram collection unsets col_en. Unsetting col_en would cause
1) Histogram interrupt handle to not stop interrupts and to
   print errors continously
2) Histogram disable to never happen so interrupts will never be
   disabled, as col_en is verified before disabling

Setting col_state to HIST_UNKNOWN can be utilized to check if the
collection was canceled. This can be verified after wait to know
if we need to send ECANCELED. If collect call arrives later it
would return EPROTO.

Change-Id: I836cfb50468c8d1e2791a7e59343f963c58cd959
Signed-off-by: Krishna Chaitanya Parimi <cparimi@codeaurora.org>